move guard test to guard.test.c
