+++ src/guard.test.c
@@ -16,6 +16,30 @@
 #include <CuTest.h>
 #include "tests.h"
 
+static void test_is_guarded(CuTest *tc) {
+    unit *u1, *u2;
+    region *r;
+    race *rc;
+
+    test_cleanup();
+    rc = rc_get_or_create("dragon");
+    rc->flags |= RCF_UNARMEDGUARD;
+    r = test_create_region(0, 0, 0);
+    u1 = test_create_unit(test_create_faction(0), r);
+    u2 = test_create_unit(test_create_faction(rc), r);
+    CuAssertPtrEquals(tc, 0, is_guarded(r, u1, GUARD_TRAVELTHRU));
+    CuAssertPtrEquals(tc, 0, is_guarded(r, u1, GUARD_PRODUCE));
+    CuAssertPtrEquals(tc, 0, is_guarded(r, u1, GUARD_TREES));
+    CuAssertPtrEquals(tc, 0, is_guarded(r, u1, GUARD_MINING));
+    guard(u2, GUARD_MINING | GUARD_PRODUCE);
+    CuAssertIntEquals(tc, GUARD_CREWS | GUARD_LANDING | GUARD_TRAVELTHRU | GUARD_TAX | GUARD_PRODUCE | GUARD_RECRUIT, guard_flags(u2));
+    CuAssertPtrEquals(tc, 0, is_guarded(r, u1, GUARD_TRAVELTHRU));
+    CuAssertPtrEquals(tc, 0, is_guarded(r, u1, GUARD_TREES));
+    CuAssertPtrEquals(tc, 0, is_guarded(r, u1, GUARD_MINING));
+    CuAssertPtrEquals(tc, u2, is_guarded(r, u1, GUARD_PRODUCE));
+    test_cleanup();
+}
+
 static void test_guard_unskilled(CuTest * tc)
 // TODO: it would be better to test armedmen()
 {
@@ -114,6 +138,7 @@ static void test_guard_monsters(CuTest * tc)
 CuSuite *get_guard_suite(void)
 {
     CuSuite *suite = CuSuiteNew();
+    SUITE_ADD_TEST(suite, test_is_guarded);
     SUITE_ADD_TEST(suite, test_is_guard);
     SUITE_ADD_TEST(suite, test_guard_unskilled);
     SUITE_ADD_TEST(suite, test_guard_armed);
